+++ test/date-tests.spec.js
@@ -16,31 +16,11 @@ import {
 
 import {Ed25519Signature2018} from '..';
 import {credential, mockKeyPair} from './mock-data.js';
-import didContext from 'did-context';
-const credentialsContext = require('credentials-context');
+import documentLoader from './documentLoader';
 
-const context = {
-  [didContext.constants.DID_CONTEXT_URL]: //
-  didContext.contexts.get(didContext.constants.DID_CONTEXT_URL),
-  [credentialsContext.constants.CONTEXT_URL]: //
-  credentialsContext.contexts.get(credentialsContext.constants.CONTEXT_URL),
-  'https://www.w3.org/2018/credentials/examples/v1': //
-  require('./contexts/example-v1.json'),
-  'https://www.w3.org/ns/odrl.jsonld': require('./contexts/odrl-v1.json')
-};
+describe('Document date behavior', () => {
 
-const documentLoader = async iri => {
-
-  if(context[iri]) {
-    return {document: context[iri]};
-  }
-
-  throw new Error('Could not find: ', iri);
-};
-
-describe('Suite date behavior', () => {
-
-  it('should sign a document when date is valid', async () => {
+  it('should sign when issuanceDate is valid', async () => {
     const unsignedCredential = {...credential};
     const keyPair = await Ed25519VerificationKey2018.from({...mockKeyPair});
     const suite = new Ed25519Signature2018({
@@ -55,8 +35,8 @@ describe('Suite date behavior', () => {
     expect(signedCredential).to.have.property('proof');
   });
 
-  it('should fail to sign a document when date is object', async () => {
-    //const ISSUED_ON = new Date("1996-09-29T15:01:23.456Z").getTime();
+  it('should catch error when issuanceDate is object', async () => {
+    // const ISSUED_ON = new Date("1996-09-29T15:01:23.456Z").getTime();
     // Produced from moment(ISSUED_ON).toObject()
     const dateObj = {
       years: 1996,
@@ -69,18 +49,26 @@ describe('Suite date behavior', () => {
     };
 
     const unsignedCredential = {...credential};
+    unsignedCredential.issuanceDate = dateObj;
+
     const keyPair = await Ed25519VerificationKey2018.from({...mockKeyPair});
     const suite = new Ed25519Signature2018({
       key: keyPair,
     });
 
-    unsignedCredential.issuanceDate = dateObj;
-    const signedCredential = await jsigs.sign(unsignedCredential, {
-      suite,
-      purpose: new AssertionProofPurpose(),
-      documentLoader,
-    });
-    expect(signedCredential).to.have.property('proof');
+    let err;
+
+    try {
+      await jsigs.sign(unsignedCredential, {
+        suite,
+        purpose: new AssertionProofPurpose(),
+        documentLoader,
+      });
+    } catch(e) {
+      err = e;
+    }
+
+    err.should.exist;
   });
 
 });
